Lista 2

+++ src/com/Unama/Ex2letraF.java
@@ -5,10 +5,10 @@ import java.util.Scanner;
 public class Ex2letraF {
     public static void main(String[] args){
         Scanner leia = new Scanner(System.in);
-        float a,b;
+        int a,b;
         System.out.println("Informe dois numeros:");
-        a = leia.nextFloat();
-        b = leia.nextFloat();
+        a = leia.nextInt();
+        b = leia.nextInt();
         if(a>b){
             System.out.print("O número maior é: "+a);
         }else {
